update sbt-unidoc in preparation for sbt 1.x
